fix: replaced mock useVoteOnProposal hook with actual implementation and fixed query error
