+++ apps/davi/src/stores/modules/1_5/fetchers/subgraph/useGetVotingMachineAddressBySchemeAddress/query.graphql
@@ -1,5 +1,7 @@
 query getVotingMachineByScheme($schemeAddress: ID!) {
   scheme(id: $schemeAddress) {
-    votingMachine
+    votingMachine {
+      id
+    }
   }
 }
